requirements.txt
Flask==3.0.3
flask-cors==4.0.0
flask-pymongo==2.3.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
pymongo==4.7.3
python-dotenv==1.0.1
email-validator==2.2.0
python-magic-bin==0.4.14; platform_system=="Windows"
python-magic==0.4.27; platform_system!="Windows"
gunicorn==22.0.0
openai==1.8.0
requests==2.31.0
python-multipart==0.0.6
boto3==1.28.0
sentry-sdk==1.22.0